@@ -1,8 +1,6 @@
-```javascript
-module.exports = {
+export default {
   plugins: {
     tailwindcss: {},
     autoprefixer: {},
   },
-}
-```
+};
